Do not test scheduled and succeeded/failed webhooks together
The design of the callback webserver is to quit once a fixed number of
callbacks have been received.  Currently, the scheduled callback may
occur multiple times while the job is pending (scheduled).

In the future we may enforce that the scheduled webhook is called
exactly once, and re-introduce these tests.